fix: checked if imageis deleted or not

+++ src/services/media.service.ts
@@ -4,6 +4,7 @@ import { Media } from '../entities/vendor/media.entity';
 import { ImageInterface } from '../interface/media.interface';
 import HttpException from '../utils/HttpException.utils';
 import { convertImageToMedia } from '../utils/imageToMedia.utils';
+import { deleteMedia } from '../utils/mediaDelete.utils';
 
 class mediaService {
     constructor(private readonly media = AppDataSource.getRepository(Media)) {}
@@ -19,16 +20,20 @@ class mediaService {
         }
     }
 
-    async deleteMedia(id: string) {
+    async deleteMedia(id: string, filepath: string) {
         try {
-            const item = await this.media
-                .createQueryBuilder('media')
-                .delete()
-                .from(Media)
-                .where('id = :id', {
-                    id: id,
-                })
-                .execute();
+            const imageDelete = await deleteMedia(filepath);
+            if (imageDelete) {
+                const item = await this.media
+                    .createQueryBuilder('media')
+                    .delete()
+                    .from(Media)
+                    .where('id = :id', {
+                        id: id,
+                    })
+                    .execute();
+                return item?.affected;
+            }
         } catch (error: any) {
             throw HttpException.badRequest(error?.message);
         }
